requirements.txt
numpy
opencv-contrib-python>=4.5.0
Pillow
pillow-heif
zxing-cpp<3
fastapi
uvicorn[standard]
python-multipart
celery[redis]>=5.3
redis>=5.0
